refactor: move Credits to /projectsV2
