+++ pages/sites/[slug]/[locale]/index.tsx
@@ -5,7 +5,7 @@ import GetAllProjectsMeta from '../../../../src/utils/getMetaTags/GetAllProjects
 import getStoredCurrency from '../../../../src/utils/countryCurrency/getStoredCurrency';
 import { getRequest } from '../../../../src/utils/apiRequests/api';
 import { useProjectProps } from '../../../../src/features/common/Layout/ProjectPropsContext';
-import Credits from '../../../../src/features/projects/components/maps/Credits';
+import Credits from '../../../../src/features/projectsV2/ProjectsMap/Credits';
 import Filters from '../../../../src/features/projects/components/projects/Filters';
 import { ErrorHandlingContext } from '../../../../src/features/common/Layout/ErrorHandlingContext';
 import DirectGift, {
