+++ src/features/common/Layout/ProjectsLayout/index.tsx
@@ -1,6 +1,6 @@
 import { FC, useMemo } from 'react';
 import styles from './ProjectsLayout.module.scss';
-import Credits from '../../../projects/components/maps/Credits';
+import Credits from '../../../projectsV2/ProjectsMap/Credits';
 import { SetState } from '../../types/common';
 import ProjectsMap from '../../../projectsV2/ProjectsMap';
 import { ProjectsProvider } from '../../../projectsV2/ProjectsContext';
